Move IndexLayerClient online test to separate file
Move online tests from TestIndexLayerClient to
IndexLayerClientOnlineTest.

Relates-to: OLPEDGE-720
