+++ olp-cpp-sdk-dataservice-write/tests/TestIndexLayerClient.cpp
@@ -145,161 +145,6 @@ class IndexLayerClientTestBase : public ::testing::TestWithParam<bool> {
   std::shared_ptr<std::vector<unsigned char>> data_;
 };
 
-class IndexLayerClientOnlineTest : public IndexLayerClientTestBase {
- protected:
-  static std::shared_ptr<olp::http::Network> s_network;
-
-  static void SetUpTestSuite() {
-    s_network = olp::client::OlpClientSettingsFactory::
-        CreateDefaultNetworkRequestHandler();
-  }
-
-  virtual std::shared_ptr<IndexLayerClient> CreateIndexLayerClient() override {
-    auto network = s_network;
-
-    olp::authentication::Settings authentication_settings;
-    authentication_settings.token_endpoint_url =
-        CustomParameters::getArgument(kEndpoint);
-    authentication_settings.network_request_handler = network;
-
-    olp::authentication::TokenProviderDefault provider(
-        CustomParameters::getArgument(kAppid),
-        CustomParameters::getArgument(kSecret), authentication_settings);
-
-    olp::client::AuthenticationSettings auth_client_settings;
-    auth_client_settings.provider = provider;
-
-    olp::client::OlpClientSettings settings;
-    settings.authentication_settings = auth_client_settings;
-    settings.network_request_handler = network;
-
-    return std::make_shared<IndexLayerClient>(
-        olp::client::HRN{GetTestCatalog()}, settings);
-  }
-};
-
-// Static network instance is necessary as it needs to outlive any created
-// clients. This is a known limitation as triggered send requests capture the
-// network instance inside the callbacks.
-std::shared_ptr<olp::http::Network> IndexLayerClientOnlineTest::s_network;
-
-INSTANTIATE_TEST_SUITE_P(TestOnline, IndexLayerClientOnlineTest,
-                         ::testing::Values(true));
-
-TEST_P(IndexLayerClientOnlineTest, PublishData) {
-  auto response = client_
-                      ->PublishIndex(PublishIndexRequest()
-                                         .WithIndex(GetTestIndex())
-                                         .WithData(data_)
-                                         .WithLayerId(GetTestLayer()))
-                      .GetFuture()
-                      .get();
-
-  ASSERT_NO_FATAL_FAILURE(PublishDataSuccessAssertions(response));
-}
-
-TEST_P(IndexLayerClientOnlineTest, DeleteData) {
-  auto response = client_
-                      ->PublishIndex(PublishIndexRequest()
-                                         .WithIndex(GetTestIndex())
-                                         .WithData(data_)
-                                         .WithLayerId(GetTestLayer()))
-                      .GetFuture()
-                      .get();
-
-  ASSERT_NO_FATAL_FAILURE(PublishDataSuccessAssertions(response));
-  auto index_id = response.GetResult().GetTraceID();
-
-  auto deleteIndexRes =
-      client_
-          ->DeleteIndexData(
-              DeleteIndexDataRequest().WithIndexId(index_id).WithLayerId(
-                  GetTestLayer()))
-          .GetFuture()
-          .get();
-
-  ASSERT_TRUE(deleteIndexRes.IsSuccessful());
-}
-
-TEST_P(IndexLayerClientOnlineTest, PublishDataAsync) {
-  std::promise<PublishIndexResponse> response_promise;
-  bool call_is_async = true;
-
-  auto cancel_token =
-      client_->PublishIndex(PublishIndexRequest()
-                                .WithIndex(GetTestIndex())
-                                .WithData(data_)
-                                .WithLayerId(GetTestLayer()),
-                            [&](const PublishIndexResponse& response) {
-                              call_is_async = false;
-                              response_promise.set_value(response);
-                            });
-
-  EXPECT_TRUE(call_is_async);
-  auto response_future = response_promise.get_future();
-  auto status = response_future.wait_for(std::chrono::seconds(30));
-  if (status != std::future_status::ready) {
-    cancel_token.cancel();
-  }
-  auto response = response_future.get();
-
-  ASSERT_NO_FATAL_FAILURE(PublishDataSuccessAssertions(response));
-}
-
-TEST_P(IndexLayerClientOnlineTest, UpdateIndex) {
-  Index index = GetTestIndex();
-  index.SetId("2f269191-5ef7-42a4-a445-fdfe53f95d92");
-
-  auto response =
-      client_
-          ->UpdateIndex(
-              UpdateIndexRequest()
-                  .WithIndexAdditions({index})
-                  .WithIndexRemovals({"2f269191-5ef7-42a4-a445-fdfe53f95d92"})
-                  .WithLayerId(GetTestLayer()))
-          .GetFuture()
-          .get();
-
-  EXPECT_TRUE(response.IsSuccessful());
-  EXPECT_EQ("", response.GetError().GetMessage());
-}
-
-TEST_P(IndexLayerClientOnlineTest, PublishNoData) {
-  auto response = client_
-                      ->PublishIndex(PublishIndexRequest()
-                                         .WithIndex(GetTestIndex())
-                                         .WithLayerId(GetTestLayer()))
-                      .GetFuture()
-                      .get();
-
-  ASSERT_NO_FATAL_FAILURE(PublishFailureAssertions(response));
-  ASSERT_EQ(olp::client::ErrorCode::InvalidArgument,
-            response.GetError().GetErrorCode());
-  ASSERT_EQ("Request data empty.", response.GetError().GetMessage());
-}
-
-TEST_P(IndexLayerClientOnlineTest, PublishNoLayer) {
-  auto response = client_
-                      ->PublishIndex(PublishIndexRequest()
-                                         .WithIndex(GetTestIndex())
-                                         .WithData(data_)
-                                         .WithLayerId("invalid-layer"))
-                      .GetFuture()
-                      .get();
-
-  ASSERT_NO_FATAL_FAILURE(PublishFailureAssertions(response));
-  ASSERT_EQ(olp::client::ErrorCode::InvalidArgument,
-            response.GetError().GetErrorCode());
-  ASSERT_EQ(
-      "Unable to find the Layer ID (invalid-layer) "
-      "provided in "
-      "the PublishIndexRequest in the "
-      "Catalog specified when creating "
-      "this "
-      "IndexLayerClient instance.",
-      response.GetError().GetMessage());
-}
-
 using ::testing::_;
 
 class IndexLayerClientMockTest : public IndexLayerClientTestBase {
